rentsel/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
pymysql==1.0.2
joblib==1.1.0
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
seaborn==0.11.2
mysql-connector-python==8.0.26
gunicorn==20.1.0